Setup basic configuration JSON file
This is where all the data is stored and saved - all the base values are written in here so if you wish to have a higher or lower budget or popularity you can change that here. Editing PCPOWER must only been done after you have bought a PC in the game since this value may be overwritten otherwise.